# packages/contracts/dist/openapi.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/redirects": {
      "get": {
        "operationId": "RedirectsController_list",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "redirects"
        ]
      },
      "post": {
        "operationId": "RedirectsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRedirectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "redirects"
        ]
      }
    },
    "/api/v1/admin/redirects/{id}": {
      "patch": {
        "operationId": "RedirectsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRedirectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "redirects"
        ]
      },
      "delete": {
        "operationId": "RedirectsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "redirects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Blueprint API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRedirectDto": {
        "type": "object",
        "properties": {
          "fromPath": {
            "type": "string"
          },
          "toPath": {
            "type": "string"
          },
          "statusCode": {
            "type": "number",
            "enum": [
              301,
              302
            ]
          }
        },
        "required": [
          "fromPath",
          "toPath",
          "statusCode"
        ]
      },
      "UpdateRedirectDto": {
        "type": "object",
        "properties": {
          "fromPath": {
            "type": "string"
          },
          "toPath": {
            "type": "string"
          },
          "statusCode": {
            "type": "number",
            "enum": [
              301,
              302
            ]
          }
        },
        "required": [
          "fromPath",
          "toPath",
          "statusCode"
        ]
      }
    }
  }
}